api/requirements.txt
flask-bcrypt==0.7.1
flask-cors==3.0.8
flask-jwt-extended==3.24.1
flask-mongoengine==0.9.5
flask-restful==0.3.7
gunicorn==20.0.4
